+++ contracts/contracts/lab/sales/holder/Factory.sol
@@ -23,29 +23,29 @@ contract HolderFactory is Auth {
 
     function deploy(
         address lab,
-        uint96 wallet,
+        uint256 wallet,
         address receiver,
-        uint96 transaction,
+        uint256 transaction,
         address token,
-        uint96 finish,
+        uint256 finish,
         address checker,
-        uint96 start,
-        uint128 supply,
-        uint128 cost,
+        uint256 start,
+        uint256 supply,
+        uint256 cost,
         uint256 item,
         address _owner
     ) external requiresAuth returns (HolderSale sale) {
         bytes memory data = abi.encodePacked(
             lab,
-            uint256(wallet),
+            wallet,
             receiver,
-            uint256(transaction),
+            transaction,
             token,
-            uint256(finish),
+            finish,
             checker,
-            uint256(start),
-            uint256(supply),
-            uint256(cost),
+            start,
+            supply,
+            cost,
             item
         );
