+++ contracts/contracts/lab/sales/publik/Data.sol
@@ -13,36 +13,36 @@ contract PublicData is Clone {
         return Lab(_getArgAddress(0));
     }
 
-    function wallet() public pure returns (uint96) {
-        return uint96(_getArgUint256(20));
+    function wallet() public pure returns (uint256) {
+        return _getArgUint256(20);
     }
 
     function receiver() public pure returns (address payable) {
         return payable(_getArgAddress(52));
     }
 
-    function transaction() public pure returns (uint96) {
-        return uint96(_getArgUint256(72));
+    function transaction() public pure returns (uint256) {
+        return _getArgUint256(72);
     }
 
     function token() public pure returns (ERC20) {
         return ERC20(_getArgAddress(104));
     }
 
-    function finish() public pure returns (uint96) {
-        return uint96(_getArgUint256(124));
+    function finish() public pure returns (uint256) {
+        return _getArgUint256(124);
     }
 
-    function start() public pure returns (uint96) {
-        return uint96(_getArgUint256(156));
+    function start() public pure returns (uint256) {
+        return _getArgUint256(156);
     }
 
-    function supply() public pure returns (uint128) {
-        return uint128(_getArgUint256(188));
+    function supply() public pure returns (uint256) {
+        return _getArgUint256(188);
     }
 
-    function cost() public pure returns (uint128) {
-        return uint128(_getArgUint256(220));
+    function cost() public pure returns (uint256) {
+        return _getArgUint256(220);
     }
 
     function item() public pure returns (uint256) {
